refactor: migrate individual serializer implementations into separate files and remove consolidated files
